cleanup network
