--- network/src/commonMain/kotlin/com/wire/kalium/network/api/user/register/RegisterResponse.kt
@@ -1,30 +1,0 @@
-package com.wire.kalium.network.api.user.register
-
-
-import com.wire.kalium.network.api.TeamId
-import com.wire.kalium.network.api.model.Service
-import com.wire.kalium.network.api.model.UserAssetDTO
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RegisterResponse(
-    @SerialName("accent_id")
-    val accentId: Int?,
-    @SerialName("assets")
-    val assets: List<UserAssetDTO>,
-    @SerialName("deleted")
-    val deleted: Boolean?,
-    @SerialName("email")
-    val email: String?,
-    @SerialName("handle")
-    val handle: String?,
-    @SerialName("id")
-    val id: String,
-    @SerialName("name")
-    val name: String,
-    @SerialName("service")
-    val service: Service?,
-    @SerialName("team")
-    val teamId: TeamId?
-)
